% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_to_process_dr_student_information.R
\name{clean_to_process_dr_student_information}
\alias{clean_to_process_dr_student_information}
\title{DR Student Information Clean to Process}
\usage{
clean_to_process_dr_student_information(district, period)
}
\arguments{
\item{raw_file_xwalk}{}
}
\description{
DR Student Information Clean to Process
}
